chore: format

+++ packages/contracts/src/deploy/deployPoolToken.ts
@@ -28,8 +28,9 @@ import {
   PublicKey,
   Signature,
   SmartContract,
-  UInt8,
-  UInt64} from "o1js"
+  UInt64,
+  UInt8
+} from "o1js"
 
 import {
   Faucet,
